Atualizando os campos de quantidade para Integer ao invés de Decimal e adicionando um contador de itens no carrinho.

+++ apps/estoque/models.py
@@ -15,7 +15,7 @@ class Material(models.Model):
 
     desc = models.CharField(max_length=100, null=False, blank=False)
     categoria = models.CharField(max_length=100, choices=OPCOES_CATEGORIA, default="")
-    qtd = models.DecimalField(max_digits=8, decimal_places=1, null=False, blank=False)
+    qtd = models.IntegerField( null=False, blank=False)
     tipo = models.CharField(max_length=100, choices=OPCOES_MEDIDA, default="UND")
     foto = models.ImageField(upload_to="fotos/%Y/%m/%d/", blank=True)
